crawler/requirements.txt
beautifulsoup4==4.12.3
selenium==4.17.2
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.0
python-dotenv==1.0.0
